+++ Package.swift
@@ -17,7 +17,7 @@ let package = Package(
         .library(name: "AlloySwiftUIControls", targets: ["AlloySwiftUIControls"]),
     ],
     dependencies: [
-        .package(url: "https://github.com/sunimp/HTTPMediaCache.git", from: "1.0.0"),
+        .package(url: "https://github.com/sunimp/HTTPMediaCache.git", from: "1.0.1"),
     ],
     targets: [
         .target(name: "AlloyCore"),